qa: move options tests to dedicated options test case
